Update README.md
I'm writing this review after watching it for the second time, I never do that for Dramas which aren't Rom coms but Something in the Rain changed that for me. I never read a full in depth review for Dramas or Movies as it hinders the watching experience but you just can't escape Socia media completely, many people reviewed this drama as boring, slow, the annoying Mom they hate passionately and talked about the ending it has, I'm glad I didn't took those reviews seriously and gave it a watch.